separating options test
