+++ bigbluebutton-tests/playwright/options/about/util.js
@@ -1,0 +1,8 @@
+const e = require('../../../core/elements');
+
+async function openAboutModal(test) {
+  await test.waitAndClick(e.optionsButton);
+  await test.waitAndClick(e.showAboutModalButton);
+}
+
+exports.openAboutModal = openAboutModal;
